fix : 트레이너 item 디자인 수정2 #1

+++ app/src/main/java/com/example/fit_i/TrainerData.kt
@@ -7,7 +7,7 @@ data class TrainerData(
     val certificate: Int,
     val univ: String,
     val pr: String,
-    val money: Double,
+    val money: Int,
     //val time: String
     //val rank : Int
 //    val img : Int
